third_party/blink/web_tests/fast/css-grid-layout/grid-item-start-before-get-set-expected.txt

Test that setting and getting grid-column-start and grid-row-start works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test getting grid-column-start and grid-row-start set through CSS
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-column') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-row') is 'auto'
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('grid-column') is "10"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('grid-row') is "15"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('grid-column') is "-10"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('grid-row') is "-15"
PASS getComputedStyle(gridItemWithBeforeSpan, '').getPropertyValue('grid-column') is "span 2"
PASS getComputedStyle(gridItemWithBeforeSpan, '').getPropertyValue('grid-row') is "span 8"
PASS getComputedStyle(gridItemWithAfterSpan, '').getPropertyValue('grid-column') is "span 2"
PASS getComputedStyle(gridItemWithAfterSpan, '').getPropertyValue('grid-row') is "span 8"
PASS getComputedStyle(gridItemWithOnlySpan, '').getPropertyValue('grid-column') is "auto"
PASS getComputedStyle(gridItemWithOnlySpan, '').getPropertyValue('grid-row') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('grid-column') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('grid-row') is "auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-column') is "first / auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-row') is "last / auto"
PASS getComputedStyle(gridItemWithNumberCustomIdent, '').getPropertyValue('grid-column') is "2 first"
PASS getComputedStyle(gridItemWithNumberCustomIdent, '').getPropertyValue('grid-row') is "3 last"
PASS getComputedStyle(gridItemWithSpanCustomIdentElement, '').getPropertyValue('grid-column') is "span first"
PASS getComputedStyle(gridItemWithSpanCustomIdentElement, '').getPropertyValue('grid-row') is "span last"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdentElement, '').getPropertyValue('grid-column') is "span 2 first"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdentElement, '').getPropertyValue('grid-row') is "span 3 last"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-column') is "firstArea / auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-row') is "thirdArea / auto"


Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto'


Test getting and setting grid-column-start and grid-row-start through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "18"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "66"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "-55"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "-40"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "nav / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "last / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 20"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "thirdArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "nonExistentArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "nonExistentArea / auto"


Test setting grid-column-start and grid-row-start to 'inherit' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "6"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "18"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "6"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 1"


Test setting grid-column-start and grid-row-start to 'initial' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "1 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 3 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 3 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "auto / 5"


Test setting grid-column-start and grid-row-start back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto'
PASS successfullyParsed is true

TEST COMPLETE
